requirements.txt


certifi==2020.4.5.2
cffi==1.12.3
chardet==3.0.4


cycler==0.10.0
Cython==0.29.13
cytoolz==0.10.0
dask==2.5.2
decorator==4.4.0
defusedxml==0.6.0
distlib==0.3.0
distributed==2.5.2
Django==3.0.3
django-crispy-forms==1.9.2
django-recaptcha==2.0.6
djangorestframework==3.12.1
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0





glob2==0.7
greenlet==0.4.15
gTTS==2.1.1
gTTS-token==1.1.3
h5py==2.9.0
HeapDict==1.0.1
html5lib==1.0.1

idna==2.8
imageio==2.6.0
imageio-ffmpeg==0.4.2
imagesize==1.1.0
importlib-metadata==0.23

ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.3
joblib==0.13.2
json5==0.8.5
jsonschema==3.0.2








lxml==4.4.1






Pillow==7.1.2
pytz==2019.3




requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
responses==0.10.15









urllib3==1.25.9
virtualenv==20.0.21

